+++ src/components/AlbumContent/CoverImage.jsx
